ML Projects/backorder-prediction/app/backorder/serializers.py
from rest_framework import serializers

from .models import ContactUs


class ContactUsSerializer(serializers.ModelSerializer):
    class Meta:
        model = ContactUs
        fields = ['name', 'email', 'message']
