ML Projects/backorder-prediction/app/requirements/common.txt
django
djangorestframework

pandas
numpy
joblib
scipy
imblearn

# [ML]
scikit-learn==1.2.2
lightgbm
xgboost
